+++ docs/examples/copy_dict.py
@@ -1,25 +1,18 @@
 from pydantic import BaseModel
 
-
 class BarModel(BaseModel):
     whatever: int
-
 
 class FooBarModel(BaseModel):
     banana: float
     foo: str
     bar: BarModel
 
-
 m = FooBarModel(banana=3.14, foo='hello', bar={'whatever': 123})
 
 print(m.dict())
 # (returns a dictionary)
 # > {'banana': 3.14, 'foo': 'hello', 'bar': {'whatever': 123}}
-
-print(m.json())
-# (returns a str)
-# > {"banana": 3.14, "foo": "hello", "bar": {"whatever": 123}}
 
 print(m.dict(include={'foo', 'bar'}))
 # > {'foo': 'hello', 'bar': {'whatever': 123}}