+++ HISTORY.rst
@@ -7,6 +7,7 @@ v0.12.2 (XXXX-XX-XX)
 ....................
 * raise an exception if a field's name shadows an existing ``BaseModel`` attribute #242
 * add ``UrlStr`` and ``urlstr`` types #236
+* timedelta json encoding ISO8601 and total seconds, custom json encoders #247, by @cfkanesan and @samuelcolvin
 
 v0.12.1 (2018-07-31)
 ....................